+++ CHANGES.rst
@@ -6,6 +6,10 @@ Changelog
 
 - Added `outputMimeType` parameter to `richtextval` method
   [sgeulette]
+- Added parameter `query={}` to `content.uuidsToCatalogBrains`, this let's you
+  complete the catalog query in case you have `UIDs` and you want to filter
+  it on additional index like `review_state`.
+  [gbastien]
 
 0.27 (2020-04-20)
 -----------------